move `AddOwnerModal` and `UserTeamSelect` out of enterprise directory to pass the "no enterprise import in OSS web guard"
